+++ src/wait/retry.rs
@@ -3,14 +3,14 @@ use std::time::Duration;
 use tokio::time::delay_for;
 use url::Url;
 
-pub async fn attempt_build_by_url(
+pub fn attempt_build_by_url(
     client: &Buildkite,
     url: &Url,
     max_attempts: u64,
 ) -> Result<Build, buildkite_waiter::error::RequestError> {
     let mut current_retry = 0u64;
 
-    let mut last_result = client.build_by_url(url.as_str()).await;
+    let mut last_result = client.build_by_url(url.as_str());
 
     while current_retry < max_attempts {
         if let Ok(build) = last_result {